# api-docs.json
{"openapi":"3.1.0","info":{"title":"PrepareForTraining Management API","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Documents","description":"API Quản lý Tài liệu & Tệp tin đính kèm của Dự án"}],"paths":{"/api/v1/projects/{projectId}/documents/{documentId}/rename":{"put":{"tags":["Documents"],"summary":"Rename a document","operationId":"rename","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"documentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenameDocumentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDocumentResponse"}}}}}}},"/api/v1/projects/{id}":{"get":{"tags":["project-controller"],"summary":"Get project details","operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}},"put":{"tags":["project-controller"],"summary":"Update a project (Owner or Admin)","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}},"delete":{"tags":["project-controller"],"summary":"Archive a project (Owner or Admin)","operationId":"archive","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}}},"/api/v1/projects/{id}/star":{"put":{"tags":["project-controller"],"summary":"Toggle star status of a project for current user","operationId":"toggleStar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}}},"/api/v1/projects/{id}/restore":{"put":{"tags":["project-controller"],"summary":"Restore an archived project (Owner or Admin)","operationId":"restore","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}}},"/api/v1/projects/{id}/pin":{"put":{"tags":["project-controller"],"summary":"Toggle pin status of a project (Owner or Admin)","operationId":"togglePin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}}},"/api/v1/projects":{"get":{"tags":["project-controller"],"summary":"List projects visible to current user","operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"filter","in":"query","required":false,"schema":{"type":"string"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt,desc"}},{"name":"search","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageProjectResponse"}}}}}},"post":{"tags":["project-controller"],"summary":"Create a project (Owner or Admin)","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectResponse"}}}}}}},"/api/v1/projects/{projectId}/documents":{"get":{"tags":["Documents"],"summary":"List documents in a project","operationId":"list_1","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDocumentResponse"}}}}}},"post":{"tags":["Documents"],"summary":"Upload a document to a project","operationId":"upload","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDocumentResponse"}}}}}}},"/api/v1/projects/{id}/members":{"get":{"tags":["project-controller"],"summary":"List project members","operationId":"members","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProjectMemberResponse"}}}}}},"post":{"tags":["project-controller"],"summary":"Add a member by email (Owner or Admin)","operationId":"addMember","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddProjectMemberRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProjectMemberResponse"}}}}}}},"/api/v1/auth/reset-password":{"post":{"tags":["auth-controller"],"summary":"Reset password using a valid token","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"summary":"Register a standard user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"summary":"Log in and receive a JWT","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/auth/google":{"post":{"tags":["auth-controller"],"summary":"Log in or register with Google","operationId":"googleLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}}}}}},"/api/v1/auth/forgot-password":{"post":{"tags":["auth-controller"],"summary":"Request a password reset link","operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/v1/admin/users":{"get":{"tags":["admin-user-controller"],"summary":"List every user","operationId":"list_2","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageUserResponse"}}}}}},"post":{"tags":["admin-user-controller"],"summary":"Create an Owner or User","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/admin/users/{id}/status":{"patch":{"tags":["admin-user-controller"],"summary":"Activate or deactivate a user","operationId":"status","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/admin/users/{id}/role":{"patch":{"tags":["admin-user-controller"],"summary":"Change a user's role","operationId":"role","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"summary":"Get profile of authenticated user","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/v1/projects/{projectId}/documents/{documentId}/download":{"get":{"tags":["Documents"],"summary":"Download a document","operationId":"download","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"documentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/v1/projects/stats":{"get":{"tags":["project-controller"],"summary":"Get dashboard stats for current user","operationId":"stats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDashboardStatsResponse"}}}}}}},"/api/v1/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/v1/projects/{projectId}/documents/{documentId}":{"delete":{"tags":["Documents"],"summary":"Delete a document","operationId":"delete","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"documentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/projects/{id}/members/{userId}":{"delete":{"tags":["project-controller"],"summary":"Remove a project member (Owner or Admin)","operationId":"removeMember","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}}},"components":{"schemas":{"RenameDocumentRequest":{"type":"object","properties":{"newName":{"type":"string"}}},"ApiResponseDocumentResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DocumentResponse"}}},"DocumentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fileName":{"type":"string"},"fileSize":{"type":"integer","format":"int64"},"contentType":{"type":"string"},"uploaderName":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"UpdateProjectRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string","maxLength":5000,"minLength":0}}},"ApiResponseProjectResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ProjectResponse"}}},"ProjectResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"ownerId":{"type":"integer","format":"int64"},"ownerName":{"type":"string"},"status":{"type":"string"},"isPinned":{"type":"boolean"},"isStarred":{"type":"boolean"},"imageCount":{"type":"integer","format":"int64"},"documentCount":{"type":"integer","format":"int64"},"videoCount":{"type":"integer","format":"int64"},"audioCount":{"type":"integer","format":"int64"}}},"CreateProjectRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string","maxLength":5000,"minLength":0}}},"AddProjectMemberRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1}}},"ApiResponseProjectMemberResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ProjectMemberResponse"}}},"ProjectMemberResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"},"joinedAt":{"type":"string","format":"date-time"}}},"ResetPasswordRequest":{"type":"object","properties":{"token":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":2147483647,"minLength":6}}},"ApiResponseString":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":72,"minLength":6},"fullName":{"type":"string","maxLength":150,"minLength":0}}},"ApiResponseAuthResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponse"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"user":{"$ref":"#/components/schemas/UserResponse"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"},"status":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"GoogleLoginRequest":{"type":"object","properties":{"accessToken":{"type":"string","minLength":1}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1}}},"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":72,"minLength":6},"fullName":{"type":"string","maxLength":150,"minLength":0},"role":{"type":"string","enum":["ADMIN","OWNER","USER"]}},"required":["role"]},"ApiResponseUserResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserResponse"}}},"UpdateUserStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["ACTIVE","INACTIVE"]}},"required":["status"]},"UpdateUserRoleRequest":{"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","OWNER","USER"]}},"required":["role"]},"ApiResponsePageProjectResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageProjectResponse"}}},"PageProjectResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"ApiResponseListDocumentResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DocumentResponse"}}}},"ApiResponseListProjectMemberResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProjectMemberResponse"}}}},"ApiResponseDashboardStatsResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DashboardStatsResponse"}}},"DashboardStatsResponse":{"type":"object","properties":{"activeProjects":{"type":"integer","format":"int64"},"totalDocuments":{"type":"integer","format":"int64"},"totalMembers":{"type":"integer","format":"int64"}}},"ApiResponsePageUserResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageUserResponse"}}},"PageUserResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}