# src/main/java/com/fpt/preparefortraining/dto/request/ForgotPasswordRequest.java
package com.fpt.preparefortraining.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record ForgotPasswordRequest(
    @NotBlank(message = "Email is required") 
    @Email(message = "Email format is invalid") 
    String email
) {}