# src/main/java/com/fpt/preparefortraining/dto/request/RenameDocumentRequest.java
package com.fpt.preparefortraining.dto.request;

public class RenameDocumentRequest {
    private String newName;

    public String getNewName() {
        return newName;
    }

    public void setNewName(String newName) {
        this.newName = newName;
    }
}