# src/main/java/com/fpt/preparefortraining/dto/request/ResetPasswordRequest.java
package com.fpt.preparefortraining.dto.request;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record ResetPasswordRequest(
    @NotBlank(message = "Token is required") 
    String token,
    
    @NotBlank(message = "New password is required") 
    @Size(min = 6, message = "Password must be at least 6 characters")
    String newPassword
) {}